+++ geotrek/feedback/models.py
@@ -13,10 +13,18 @@ from mapentity.models import MapEntityMixin
 
 from geotrek.common.mixins import TimeStampedModelMixin
 
-from .helpers import send_report_managers
+from .helpers import send_report_managers, post_report_to_suricate
 
 
 logger = logging.getLogger(__name__)
+
+
+def status_default():
+    """Set status to New by default"""
+    new_status_query = ReportStatus.objects.filter(label='Nouveau')
+    if new_status_query:
+        return new_status_query.get().pk
+    return None
 
 
 class Report(MapEntityMixin, PicturesMixin, TimeStampedModelMixin):
@@ -26,17 +34,26 @@ class Report(MapEntityMixin, PicturesMixin, TimeStampedModelMixin):
     comment = models.TextField(blank=True,
                                default="",
                                verbose_name=_("Comment"))
+    activity = models.ForeignKey('ReportActivity',
+                                 on_delete=models.CASCADE,
+                                 null=True,
+                                 blank=True,
+                                 verbose_name=_("Activity"))
     category = models.ForeignKey('ReportCategory',
                                  on_delete=models.CASCADE,
                                  null=True,
                                  blank=True,
-                                 default=None,
                                  verbose_name=_("Category"))
+    problem_magnitude = models.ForeignKey('ReportProblemMagnitude',
+                                          null=True,
+                                          blank=True,
+                                          on_delete=models.CASCADE,
+                                          verbose_name=_("Problem magnitude"))
     status = models.ForeignKey('ReportStatus',
                                on_delete=models.CASCADE,
                                null=True,
                                blank=True,
-                               default=None,
+                               default=status_default,
                                verbose_name=_("Status"))
     geom = models.PointField(null=True,
                              blank=True,
@@ -65,7 +82,7 @@ class Report(MapEntityMixin, PicturesMixin, TimeStampedModelMixin):
         return "Anonymous report"
 
     @property
-    def name_display(self):
+    def email_display(self):
         return '<a data-pk="%s" href="%s" title="%s" >%s</a>' % (self.pk,
                                                                  self.get_detail_url(),
                                                                  self,
@@ -96,27 +113,71 @@ def on_report_saved(sender, instance, created, **kwargs):
         logger.error('Email could not be sent to managers.')
         logger.exception(e)  # This sends an email to admins :)
 
+    if settings.SURICATE_REPORT_ENABLED:
+        try:
+            post_report_to_suricate(instance)
+        except Exception as e:
+            logger.error('Report could not be sent to Suricate API.')
+            logger.exception(e)
+
+
+class ReportActivity(models.Model):
+    """Activity involved in report"""
+    label = models.CharField(verbose_name=_("Activity"),
+                             max_length=128)
+    suricate_id = models.PositiveIntegerField(_("Suricate id"),
+                                              null=True,
+                                              blank=True)
+
+    class Meta:
+        verbose_name = _("Activity")
+        verbose_name_plural = _("Activities")
+        ordering = ("label",)
+
+    def __str__(self):
+        return self.label
+
 
 class ReportCategory(models.Model):
-    category = models.CharField(verbose_name=_("Category"),
-                                max_length=128)
+    label = models.CharField(verbose_name=_("Category"),
+                             max_length=128)
+    suricate_id = models.PositiveIntegerField(_("Suricate id"),
+                                              null=True,
+                                              blank=True)
 
     class Meta:
         verbose_name = _("Category")
         verbose_name_plural = _("Categories")
-        ordering = ("category",)
+        ordering = ("label",)
 
     def __str__(self):
-        return self.category
+        return self.label
 
 
 class ReportStatus(models.Model):
-    status = models.CharField(verbose_name=_("Status"),
-                              max_length=128)
+    label = models.CharField(verbose_name=_("Status"),
+                             max_length=128)
 
     class Meta:
         verbose_name = _("Status")
         verbose_name_plural = _("Status")
 
     def __str__(self):
-        return self.status
+        return self.label
+
+
+class ReportProblemMagnitude(models.Model):
+    """Report problem magnitude"""
+    label = models.CharField(verbose_name=_("Problem magnitude"),
+                             max_length=128)
+    suricate_id = models.PositiveIntegerField(_("Suricate id"),
+                                              null=True,
+                                              blank=True)
+
+    class Meta:
+        verbose_name = _("Problem magnitude")
+        verbose_name_plural = _("Problem magnitudes")
+        ordering = ("id",)
+
+    def __str__(self):
+        return self.label