+++ geotrek/feedback/tests/test_views.py
@@ -7,9 +7,23 @@ from django.core import mail
 from mapentity.factories import SuperUserFactory, UserFactory
 
 from geotrek.common.tests import CommonTest, TranslationResetMixin
+from geotrek.common.utils.testdata import get_dummy_uploaded_image_svg, get_dummy_uploaded_image, get_dummy_uploaded_file
 from geotrek.feedback import models as feedback_models
 from geotrek.feedback import factories as feedback_factories
 from rest_framework.test import APIClient
+
+
+class ReportModelTest(TestCase):
+    """Test some custom model"""
+
+    def test_default_no_status(self):
+        my_report = feedback_factories.ReportFactory()
+        self.assertEqual(my_report.status, None)
+
+    def test_default_status_exists(self):
+        self.default_status = feedback_factories.ReportStatusFactory(label="Nouveau")
+        my_report = feedback_factories.ReportFactory()
+        self.assertEqual(my_report.status, self.default_status)
 
 
 class ReportViewsetMailSend(TestCase):
@@ -19,8 +33,9 @@ class ReportViewsetMailSend(TestCase):
             {
                 'email': 'test@geotrek.local',
                 'comment': 'Test comment',
+                'activity': feedback_factories.ReportActivityFactory.create().pk,
+                'problem_magnitude': feedback_factories.ReportProblemMagnitudeFactory.create().pk,
             })
-
         self.assertEqual(len(mail.outbox), 2)
         self.assertEqual(mail.outbox[1].subject, "Geotrek : Signal a mistake")
         self.assertIn("We acknowledge receipt of your feedback", mail.outbox[1].body)
@@ -38,11 +53,13 @@ class ReportViewsTest(CommonTest):
 
     def get_expected_json_attrs(self):
         return {
-            'category': None,
+            'activity': self.obj.activity.pk,
+            'category': self.obj.category.pk,
             'comment': self.obj.comment,
             'context_object': None,
             'email': self.obj.email,
             'status': None,
+            'problem_magnitude': self.obj.problem_magnitude.pk
         }
 
     def get_bad_data(self):
@@ -52,6 +69,8 @@ class ReportViewsTest(CommonTest):
         return {
             'geom': '{"type": "Point", "coordinates": [0, 0]}',
             'email': 'yeah@you.com',
+            'activity': feedback_factories.ReportActivityFactory.create().pk,
+            'problem_magnitude': feedback_factories.ReportProblemMagnitudeFactory.create().pk,
         }
 
     def test_good_data_with_name(self):
@@ -92,6 +111,8 @@ class CreateReportsAPITest(BaseAPITest):
         self.data = {
             'geom': '{"type": "Point", "coordinates": [3, 46.5]}',
             'email': 'yeah@you.com',
+            'activity': feedback_factories.ReportActivityFactory.create().pk,
+            'problem_magnitude': feedback_factories.ReportProblemMagnitudeFactory.create().pk,
         }
 
     def post_report_data(self, data):
@@ -112,21 +133,57 @@ class CreateReportsAPITest(BaseAPITest):
         self.post_report_data(self.data)
         self.assertTrue(feedback_models.Report.objects.filter(email='yeah@you.com').exists())
 
+    def test_reports_with_file(self):
+        self.data['file'] = get_dummy_uploaded_file()
+        self.data['csv'] = get_dummy_uploaded_image_svg()
+        self.data['image'] = get_dummy_uploaded_image()
+        self.post_report_data(self.data)
+        self.assertTrue(feedback_models.Report.objects.filter(email='yeah@you.com').exists())
+        report = feedback_models.Report.objects.get()
+        self.assertEqual(report.attachments.count(), 3)
+
 
 class ListCategoriesTest(TranslationResetMixin, BaseAPITest):
     def setUp(self):
         super(ListCategoriesTest, self).setUp()
-        self.cat = feedback_factories.ReportCategoryFactory(category_it='Obstaculo')
+        self.cat = feedback_factories.ReportCategoryFactory(label_it='Obstaculo')
 
     def test_categories_can_be_obtained_as_json(self):
         response = self.client.get('/api/en/feedback/categories.json')
         self.assertEqual(response.status_code, 200)
         data = response.json()
         self.assertEqual(data[0]['id'], self.cat.id)
-        self.assertEqual(data[0]['label'], self.cat.category)
+        self.assertEqual(data[0]['label'], self.cat.label)
 
     def test_categories_are_translated(self):
         response = self.client.get('/api/it/feedback/categories.json')
         self.assertEqual(response.status_code, 200)
         data = response.json()
-        self.assertEqual(data[0]['label'], self.cat.category_it)
+        self.assertEqual(data[0]['label'], self.cat.label_it)
+
+
+class ListOptionsTest(TranslationResetMixin, BaseAPITest):
+    def setUp(self):
+        super(ListOptionsTest, self).setUp()
+        self.activity = feedback_factories.ReportActivityFactory(label_it='Hiking')
+        self.cat = feedback_factories.ReportCategoryFactory(label_it='Obstaculo')
+        self.pb_magnitude = feedback_factories.ReportProblemMagnitudeFactory(label_it='Possible')
+
+    def test_options_can_be_obtained_as_json(self):
+        response = self.client.get('/api/en/feedback/options.json')
+        self.assertEqual(response.status_code, 200)
+        data = response.json()
+        self.assertEqual(data['activities'][0]['id'], self.activity.id)
+        self.assertEqual(data['activities'][0]['label'], self.activity.label)
+        self.assertEqual(data['categories'][0]['id'], self.cat.id)
+        self.assertEqual(data['categories'][0]['label'], self.cat.label)
+        self.assertEqual(data['magnitudeProblems'][0]['id'], self.pb_magnitude.id)
+        self.assertEqual(data['magnitudeProblems'][0]['label'], self.pb_magnitude.label)
+
+    def test_options_are_translated(self):
+        response = self.client.get('/api/it/feedback/options.json')
+        self.assertEqual(response.status_code, 200)
+        data = response.json()
+        self.assertEqual(data['activities'][0]['label'], self.activity.label_it)
+        self.assertEqual(data['categories'][0]['label'], self.cat.label_it)
+        self.assertEqual(data['magnitudeProblems'][0]['label'], self.pb_magnitude.label_it)