+++ geotrek/feedback/urls.py
@@ -5,13 +5,14 @@ from mapentity.registry import registry
 from geotrek.feedback import models as feedback_models
 from geotrek.common.urls import LangConverter
 
-from .views import CategoryList
+from .views import CategoryList, FeedbackOptionsView
 
 register_converter(LangConverter, 'lang')
 
 app_name = 'feedback'
 urlpatterns = [
     path('api/<lang:lang>/feedback/categories.json', CategoryList.as_view(), name="categories_json"),
+    path('api/<lang:lang>/feedback/options.json', FeedbackOptionsView.as_view(), name="options_json"),
 ]
 
 urlpatterns += registry.register(feedback_models.Report, menu=settings.REPORT_MODEL_ENABLED)