+++ geotrek/settings/base.py
@@ -763,6 +763,17 @@ ONLY_EXTERNAL_PUBLIC_PDF = False
 
 SEND_REPORT_ACK = True
 
+SURICATE_REPORT_ENABLED = False
+
+SURICATE_REPORT_SETTINGS = {
+    'URL': '',
+    'ID_ORIGIN': '',
+    'PRIVATE_KEY_CLIENT_SERVER': '',
+    'PRIVATE_KEY_SERVER_CLIENT': '',
+}
+
+REPORT_FILETYPE = "Report"
+
 # Parser parameters for retries and error codes
 PARSER_RETRY_SLEEP_TIME = 60  # time of sleep between requests
 PARSER_NUMBER_OF_TRIES = 3  # number of requests to try before abandon